move `LinecutGenerator` and `linecut` to a dedicated module
